requirements.txt
absl-py==0.8.1
alabaster==0.7.12
astor==0.8.0
astroid==2.3.3
attrs==19.3.0
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
cloudpickle==1.2.2
colorama==0.4.1
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
imagesize==1.1.0
importlib-metadata==1.1.0
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==19.3.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.0.0
nbconvert==5.6.1
nbformat==4.4.0
numpy==1.17.4
numpydoc==0.9.1
oauthlib==3.1.0
opt-einsum==3.1.0
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
prompt-toolkit==3.0.2
protobuf==3.11.1
psutil==5.6.7
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
pyparsing==2.4.5
PyQt5==5.12.3
PyQt5-sip==12.7.0
PyQtWebEngine==5.12.1
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
PyYAML==5.2
pyzmq==18.1.1
QtAwesome==0.6.0
qtconsole==4.6.0
QtPy==1.9.0
requests==2.22.0
requests-oauthlib==1.3.0
rope==0.14.0
rsa==4.0
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.2.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==3.3.6
spyder-kernels==0.5.2
tensorboard==2.0.2
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
termcolor==1.1.0
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
